requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-text-splitters
chromadb
pypdf
docx2txt
python-pptx
openpyxl
python-dotenv
fastapi
python-multipart
uvicorn[standard]
beautifulsoup4
feedparser
ddgs
requests
websockets
ruff
